+++ LanguageExt.Core/DataTypes/Range/BigIntegerRange.cs
@@ -3,9 +3,9 @@ using System.Numerics;
 
 namespace LanguageExt
 {
-    public class BigIntegerRange : Range<BigIntegerRange, TBigInt, BigInteger>
+    public class BigIntegerRange : Range<BigIntegerRange, TBigInt, bigint>
     {
-        BigIntegerRange(BigInteger min, BigInteger max) : base(min, max, 1) { }
-        BigIntegerRange(BigInteger min, BigInteger max, BigInteger step) : base(min, max, step) { }
+        BigIntegerRange(bigint min, bigint max) : base(min, max, 1) { }
+        BigIntegerRange(bigint min, bigint max, bigint step) : base(min, max, step) { }
     }
 }
